+++ whatsnew.rst
@@ -3,10 +3,15 @@ What's New
 
 These are new features and improvements of note in each release.
 
+v0.4.0
+------
+
+ - configurations returned by :py:func:`implactlab_tools.utils.files.get_file_config`, :py:func:`implactlab_tools.utils.files.get_argv_config`, and :py:func:`implactlab_tools.utils.files.get_allargv_config` are now objects of a new :py:class:`impactlab_tools.utils.configdict.ConfigDict` class (:issue:`434`)
+
 v0.3.1 (March 19, 2018)
 -----------------------
 
- - Stability and performance improvements.
+ - "stability and performance improvements"
 
 v0.3.0 (March 17, 2018)
 -----------------------